requirements.txt
# GNSS.AI Point Collector - Python Dependencies

# Core
pyserial>=3.5
Flask>=2.3.0
Flask-SocketIO>=5.3.0
simple-websocket>=0.10.0

# Optional: Bluetooth (requires libbluetooth-dev)
# pybluez>=0.23

# Optional: ML/Data Analysis
# numpy>=1.24.0
# pandas>=2.0.0
# scikit-learn>=1.3.0
# joblib>=1.3.0

# Optional: Image processing for icons
# Pillow>=10.0.0

# Development
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
